refactored PriorityOffset to an interface
