+++ javaagents/src/main/java/g6Agent/communicationModule/submodules/PingListener.java
@@ -1,14 +1,13 @@
 package g6Agent.communicationModule.submodules;
 
 import eis.iilang.Percept;
-import g6Agent.communicationModule.CommunicationModule;
-import g6Agent.goals.PingReceiver;
+import g6Agent.goals.interfaces.PingReceiver;
 import g6Agent.services.Point;
 
 public interface PingListener {
-    public void setPingReceiver(PingReceiver pingReceiver);
+    void setPingReceiver(PingReceiver pingReceiver);
 
-    public void sendPing(Point ping);
+    void sendPing(Point ping);
 
-    public void receivePing(Percept message, String sender);
+    void receivePing(Percept message, String sender);
 }
