+++ javaagents/src/main/java/g6Agent/goals/interfaces/PingReceiver.java
@@ -1,10 +1,10 @@
-package g6Agent.goals;
+package g6Agent.goals.interfaces;
 
 import g6Agent.communicationModule.submodules.PingListener;
 import g6Agent.services.Point;
 
 //enables a goal to send and receive pings
 public interface PingReceiver {
-    public void addPingListener(PingListener pingListener);
-    public void recievePing(Point ping);
+    void addPingListener(PingListener pingListener);
+    void recievePing(Point ping);
 }
